services/order-management/requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.25.0
pydantic==2.5.3
kafka-python==2.0.2
redis==5.0.1
requests==2.31.0
kiteconnect==4.2.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
prometheus-client==0.19.0
aiohttp==3.9.1
